products/serializers.py
from rest_framework import serializers
from products.models import Category, Product
from django.utils.text import slugify


class ProductSerializer(serializers.ModelSerializer):
    class Meta:
        model = Product
        fields = ['id', 'name', 'quantity', 'price', 'category']
    
    def create(self, validated_data):
        product = Product(**validated_data)
        product.slug = slugify(product.name)
        product.save()
        return product