requirements.txt
asgiref==3.6.0
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
cryptography==39.0.0
defusedxml==0.7.1
dj-database-url==1.2.0
Django==4.2.11
django-allauth==0.52.0
django-cache-url==3.4.4
django-ckeditor==6.5.1
django-contrib-comments==2.2.0
django-crispy-forms==1.14.0
django-jalali-date==1.1.0
django-js-asset==2.0.0
django-rosetta==0.10.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
environs==9.5.0
exceptiongroup==1.1.0
factory-boy==3.3.0
Faker==20.0.3
idna==3.4
iniconfig==2.0.0
jalali_core==1.0.0
jdatetime==5.0.0
marshmallow==3.19.0
oauthlib==3.2.2
packaging==22.0
Pillow==9.3.0
pluggy==1.0.0
polib==1.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.6.0
pytest==7.2.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
setuptools==75.1.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
urllib3==1.26.13
gewheel==0.44.0
